Preview

+++ Fonts Downloader/Form1.cs
@@ -43,12 +43,12 @@ namespace Fonts_Downloader
             {
                 FolderName = folderBrowserDialog.SelectedPath;
             }
-            textBox1.Text = FolderName;
+            SelectedFolder.Text = FolderName;
         }
 
         private void FontBox1_SelectionChangeCommitted(object sender, EventArgs e)
         {
-            Key = textBox1.Text;
+            Key = ApiKeyBox.Text;
             Size.SizeStylesLoad(FontBox1, SelectedFont, SizeAndStyle, Items, FolderName);
             Document.CreateHtml(SelectedFont, SizeAndStyle, FolderName);
             //string path = System.IO.Directory.GetParent(System.IO.Directory.GetParent(Environment.CurrentDirectory).ToString()).ToString();
@@ -60,8 +60,11 @@ namespace Fonts_Downloader
 
         private void textBox2_TextChanged(object sender, EventArgs e)
         {
-            Key= textBox2.Text;
-            _ = Res.resAsync(FontBox1, Key);
+            Key = ApiKeyBox.Text;
+            if (!string.IsNullOrEmpty(Key))
+            {
+                _ = Res.resAsync(FontBox1, Key);
+            }
         }
 
         private void Form1_FormClosed(object sender, FormClosedEventArgs e)
@@ -77,29 +80,36 @@ namespace Fonts_Downloader
         }
 
         private void CopyFont_Click(object sender, EventArgs e)
-        {          
-            //webView21.CoreWebView2.NavigateToString(reader.ReadToEnd());
-            string[] FontFileStyles = { "Thin", "ExtraLight", "Light", "Regular", "Medium", "SemiBold", "Bold", "ExtraBold", "Black" };
-            var FontWeight = new List<string>();
-            var css = new CSS();
-            var File = new FontFiles();
-            if (FontBox1.SelectedItem != null)
+        {
+            if (!string.IsNullOrEmpty(FolderName))
             {
-                string FontName = SelectedFont.Text;
-                Directory.CreateDirectory($"{FolderName}\\{FontName.Replace(" ", "")}");
-                DirectoryInfo di = new DirectoryInfo($"{FolderName}\\{FontName.Replace(" ", "")}");
-                foreach (FileInfo file in di.EnumerateFiles())
+                string[] FontFileStyles = { "Thin", "ExtraLight", "Light", "Regular", "Medium", "SemiBold", "Bold", "ExtraBold", "Black" };
+                var FontWeight = new List<string>();
+                var css = new CSS();
+                var File = new FontFiles();
+                if (FontBox1.SelectedItem != null)
                 {
-                    if (file.Name.Contains(".ttf"))
+                    string FontName = SelectedFont.Text;
+                    Directory.CreateDirectory($"{FolderName}\\{FontName.Replace(" ", "")}");
+                    DirectoryInfo di = new DirectoryInfo($"{FolderName}\\{FontName.Replace(" ", "")}");
+                    foreach (FileInfo file in di.EnumerateFiles())
                     {
-                        file.Delete();
+                        if (file.Name.Contains(".ttf"))
+                        {
+                            file.Delete();
+                        }
                     }
+                    css.CreateCSS(SizeAndStyle, SubSets, FolderName, FontName, FontFileStyle);
+                    FontWeight = css.FontWeight;
+                    File.FileLinks(SizeAndStyle, FontStyle, FontWeight, FontFileStyle, FontFileStyles, SelectedFont, FolderName, FontName, Res);
+
                 }
-                css.CreateCSS(SizeAndStyle, SubSets, FolderName, FontName, FontFileStyle);
-                FontWeight = css.FontWeight;
-                File.FileLinks(SizeAndStyle, FontStyle, FontWeight, FontFileStyle, FontFileStyles, SelectedFont, FolderName, FontName, Res);
-               
             }
+            else
+            {
+                MessageBox.Show("Please select a folder ");
+            }
+
         }
         private void webView21_CoreWebView2InitializationCompleted(object sender, Microsoft.Web.WebView2.Core.CoreWebView2InitializationCompletedEventArgs e)
         {
